# packages/server/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/serverless/call": {
            "get": {
                "operationId": "ServerlessController_call_get",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "post": {
                "operationId": "ServerlessController_call_post",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "put": {
                "operationId": "ServerlessController_call_put",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "delete": {
                "operationId": "ServerlessController_call_delete",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "patch": {
                "operationId": "ServerlessController_call_patch",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "options": {
                "operationId": "ServerlessController_call_options",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "head": {
                "operationId": "ServerlessController_call_head",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            },
            "search": {
                "operationId": "ServerlessController_call_search",
                "parameters": [
                    {
                        "name": "name",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Serverless"
                ]
            }
        },
...
    },
    "info": {
        "title": "Best Swagger",
        "description": "The best API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
...
}